+++ src/forestutils.h
@@ -49,6 +49,8 @@ NumericVector cohortFoliarBiomass(List x, DataFrame SpParams, double gdd = NA_RE
 NumericVector cohortEquilibriumLeafLitter(List x, DataFrame SpParams, double AET = 800, String mode = "MED");
 NumericVector cohortEquilibriumSmallBranchLitter(List x, DataFrame SpParams, double smallBranchDecompositionRate = 0.81, String mode = "MED");
 
+NumericVector cohortCover(List x, DataFrame SpParams, String mode = "MED");
+
 NumericVector treeLAI(IntegerVector SP, NumericVector N, NumericVector dbh, DataFrame SpParams, NumericVector pEmb=NumericVector(0), double gdd = NA_REAL);
 NumericVector shrubLAI(IntegerVector SP, NumericVector Cover, NumericVector H, DataFrame SpParams, double gdd = NA_REAL);
 NumericVector cohortLAI(List x, DataFrame SpParams, double gdd = NA_REAL, String mode = "MED");
@@ -64,3 +66,4 @@ void deleteTreeCohort(List x, int treeCohort);
 void deleteShrubCohort(List x, int shrubCohort);
 
 int minDBHTreeCohort(List x, double excludeMin = 0.0);
+